backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioredis==1.3.1
amqp==2.6.1
apscheduler==3.8.1
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
automat==20.2.0
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
django-celery-results==1.2.1
django-channels==0.7.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
django-twisted==0.1.0
django==3.1.3
djangorestframework-simplejwt==4.4.0
djangorestframework==3.12.2
djoser==2.1.0
drf-spectacular==0.11.0
drf-yasg==1.20.0
gunicorn==20.0.4
hiredis==1.1.0
httpie==2.3.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
jinja2==2.11.2
jsonschema==3.2.0
kafka-python==2.0.2
kombu==4.6.11
lazy-object-proxy==1.4.3
markdown==3.3.3
markupsafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
oauthlib==3.1.0
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prompt-toolkit==3.0.8
psycopg2-binary==2.8.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pygments==2.7.2
pyhamcrest==2.0.2
pyjwt==1.7.1
pylint==2.6.0
pyopenssl==19.1.0
pyparsing==2.4.7
pyroute2==0.5.14
pyrsistent==0.17.3
pysocks==1.7.1
python-docx==0.8.11
python-dotenv==0.15.0
python-nmap==0.6.1
python3-openid==3.2.0
pytz==2020.4
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests[socks]==2.24.0
ruamel.yaml.clib==0.2.2
ruamel.yaml==0.16.12
scapy==2.4.4
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
aspose-words==23.2.0
elasticsearch==8.6.2
paramiko==2.7.2
redis